# vocal-separator/swagger-spec/swagger.yaml
openapi: 3.0.0
info:
  title: Vocal Separator API
  description: |
    Separate vocals from accompaniment in uploaded audio files, using Demucs for
    source separation with optional volume normalization, playback-speed
    adjustment, format conversion, and streaming preview.

    **Processing model**: `/api/upload` queues a Celery background task and
    returns immediately with a `task_id`. Poll `/api/status/{task_id}` until
    `status` is `SUCCESS` (or `FAILURE`), then use `task_id` with the
    download/export/preview endpoints.

    **Speed adjustment model**: `/api/process/{task_id}` queues a *separate*
    background job and returns a *new* `task_id` for it (distinct from the
    path `task_id`, which identifies the already-separated source). Poll
    `/api/status/{that new task_id}` until `SUCCESS`, then download with
    `/api/download/{original task_id}/{vocals|accompaniment}?speed={speed}`.
    The original 1.0x stems are never modified — each requested speed is
    cached and stored alongside them, so multiple speeds can coexist and a
    repeat request for the same speed returns instantly from cache.
  version: 1.0.1
  contact:
    name: Vocal Separator Team
    url: https://github.com/vocal-separator
servers:
  - url: http://localhost:8000
    description: Local development server (docker compose, port 8000)
...
paths:
...
  /api/upload:
    post:
      summary: Upload Audio File
      description: |
        Uploads an audio file and queues it for vocal/accompaniment separation
        via a Celery background task. Returns immediately (HTTP 202) with a
        `task_id` — actual separation happens asynchronously; poll
        `/api/status/{task_id}` for progress.
      tags: [Upload]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: "Audio file. Allowed extensions: .mp3, .wav, .flac, .ogg, .m4a"
                separation_intensity:
                  type: number
                  format: float
                  default: 0.5
                  minimum: 0.0
                  maximum: 1.0
                  description: Separation intensity passed to the Demucs task
                normalize:
                  type: boolean
                  default: true
                  description: Apply volume normalization to the separated stems
                speed:
                  type: number
                  format: float
                  default: 1.0
                  enum: [0.5, 0.75, 1.0, 1.25, 1.5, 2.0]
                  description: Speed factor forwarded to the separation task
                normalization_method:
                  type: string
                  default: peak
                  enum: [peak, lufs]
                  description: "'peak' targets -1 dBFS (safe headroom); 'lufs' targets -14 LUFS (YouTube/Spotify loudness target). Ignored if normalize=false."
      responses:
        '202':
          description: File accepted and queued for background processing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResponse'
              example:
                task_id: "d1bf7974-cb1e-4bbd-948b-7f547567691f"
                status: PENDING
                message: Audio uploaded and queued for separation
        '400':
          description: Unsupported file extension, or separation_intensity/speed out of range
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '413':
          description: File exceeds the configured size limit (default 500 MB)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '422':
          description: Missing required `file` field or malformed multipart body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '500':
          description: Server could not queue the processing task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
  /api/youtube:
    post:
      summary: Extract Audio from YouTube and Start Separation
      description: |
        Downloads the audio-only stream from a YouTube video (never the
        video track), encodes it to 320kbps MP3, then runs it through the
        exact same separation pipeline as `/api/upload` - same task_id
        shape, same polling via `/api/status/{task_id}`, same waveform/
        volume/speed features.

        Video duration is checked from metadata **before** any download
        starts (max 15 minutes) - an oversized or invalid video is
        rejected immediately with no Celery job ever queued.

        Progress has two stages, both visible via `/api/status/{task_id}`'s
        `stage` field: `"downloading"` (0-30%) then `"separating"` (30-100%).
      tags: [YouTube]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url]
              properties:
                url:
                  type: string
                  description: "A youtube.com/watch, youtu.be, or /shorts/ URL. Any other domain is rejected."
                  example: "https://www.youtube.com/watch?v=y6120QOlsfU"
                separation_intensity: { type: number, format: float, default: 0.5, minimum: 0.0, maximum: 1.0 }
                normalize: { type: boolean, default: true }
                normalization_method: { type: string, enum: [peak, lufs], default: peak }
                speed: { type: number, format: float, default: 1.0, enum: [0.5, 0.75, 1.0, 1.25, 1.5, 2.0] }
      responses:
        '202':
          description: Video accepted; download and separation queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResponse'
              example:
                task_id: "5819d377-a84d-4d24-9850-1eb8e37ceefd"
                status: PENDING
                message: "Downloading 'Darude - Sandstorm' (232s) and queued for separation"
        '400':
          description: Not a YouTube URL, video exceeds the 15-minute limit, or invalid processing params
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
              examples:
                not_youtube:
                  value:
                    detail: Not a supported YouTube URL
                too_long:
                  value:
                    detail: "Video is 20.0 min long. Maximum allowed: 15 min."
        '404':
          description: Video unavailable (private, deleted, geo-restricted, age-restricted, or a livestream)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
  /api/status/{task_id}:
    get:
      summary: Get Processing Status
      description: |
        Reads task state directly from Celery's `AsyncResult`. Note: an
        unknown/nonexistent `task_id` returns HTTP 200 with `status: PENDING`
        rather than 404 — this is Celery's default behavior for unseen task IDs,
        not a lookup against a persisted task table.
      tags: [Task Status]
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Current task status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
              examples:
                pending:
                  value:
                    task_id: "37f60a94-fb76-4cba-aacc-5e708092ebb4"
                    status: PENDING
                    progress: 0
                    vocals_url: null
                    accompaniment_url: null
                    error: null
                processing:
                  value:
                    task_id: "37f60a94-fb76-4cba-aacc-5e708092ebb4"
                    status: PROCESSING
                    progress: 57
                    vocals_url: null
                    accompaniment_url: null
                    error: null
                success:
                  value:
                    task_id: "37f60a94-fb76-4cba-aacc-5e708092ebb4"
                    status: SUCCESS
                    progress: 100
                    vocals_url: /api/download/37f60a94-fb76-4cba-aacc-5e708092ebb4/vocals
                    accompaniment_url: /api/download/37f60a94-fb76-4cba-aacc-5e708092ebb4/accompaniment
                    error: null
                failure:
                  value:
                    task_id: "4fe8943f-4a04-4e89-9d44-74e4b14c3954"
                    status: FAILURE
                    progress: 0
                    vocals_url: null
                    accompaniment_url: null
                    error: "Command 'ffprobe ...' returned non-zero exit status 1."
  /api/download/{task_id}/{file_type}:
    get:
      summary: Download Separated Audio
      description: |
        Downloads a completed stem as an attachment, in whatever format the
        server produced it in (MP3 by default, controlled by the
        `AUDIO_OUTPUT_FORMAT` env var, with automatic WAV/MP3 fallback if the
        primary format is missing on disk).
      tags: [Download]
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
        - name: file_type
          in: path
          required: true
          schema:
            type: string
            enum: [vocals, accompaniment, both, original]
          description: "\"both\" is a real vocals+accompaniment mixdown; \"original\" is a compressed copy of the source kept for comparison even though the raw upload is deleted after separation. Neither supports the `speed` param below."
        - name: speed
          in: query
          required: false
          schema:
            type: number
            format: float
          description: |
            Optional. When set to a value other than 1.0, downloads the
            speed-adjusted variant previously produced by
            `POST /api/process/{task_id}?speed={speed}` instead of the
            1.0x master. Must match a speed that was already requested and
            completed via that endpoint, or this returns 404.
      responses:
        '200':
          description: Audio file attachment (master, or the requested speed variant)
          content:
            audio/mpeg:
              schema:
                type: string
                format: binary
            audio/wav:
              schema:
                type: string
                format: binary
        '404':
          description: Output not found for this task_id/file_type, or the requested speed variant hasn't been produced yet
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
              examples:
                no_output:
                  value:
                    detail: Audio output not found
                no_speed_variant:
                  value:
                    detail: "No 1.5x variant found. Call POST /api/process/37f60a94-fb76-4cba-aacc-5e708092ebb4?speed=1.5 first."
        '422':
          description: file_type is not "vocals" or "accompaniment"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
  /api/waveform/{task_id}:
    get:
      summary: Get Waveform Comparison Data
      description: |
        Returns downsampled RMS peaks (800 points, independent of track
        length) for the original upload, the separated vocals, and the
        separated accompaniment — for a visual before/after comparison.

        Generated once during separation (before the original upload is
        deleted) and cached to `waveforms.json` inside the task's output
        folder; this endpoint just reads that file, so it's only available
        once the task has reached SUCCESS.
      tags: [Waveform]
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Waveform peaks for all three tracks
          content:
            application/json:
              schema:
                type: object
                properties:
                  task_id: { type: string }
                  num_points: { type: integer, example: 800 }
                  original:
                    type: array
                    items: { type: number }
                  vocals:
                    type: array
                    items: { type: number }
                  accompaniment:
                    type: array
                    items: { type: number }
                  both:
                    type: array
                    items: { type: number }
        '404':
          description: Waveform data not found (still processing, failed, or predates this feature)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
...
components:
  schemas:
    UploadResponse:
      type: object
      properties:
        task_id: { type: string }
        status: { type: string, enum: [PENDING] }
        message: { type: string }
...
    ErrorDetail:
      type: object
      properties:
        detail:
          type: string
    ValidationError:
      description: Standard FastAPI/Pydantic 422 shape
      type: object
      properties:
        detail:
          type: array
          items:
            type: object
            properties:
              type: { type: string }
              loc:
                type: array
                items: {}
              msg: { type: string }
              input: {}